test(helpers): Move helpers to test/helpers folder
