+++ test/dxdao.js
@@ -1,5 +1,5 @@
 import * as helpers from "./helpers";
-const constants = require("./constants");
+const constants = require("./helpers/constants");
 const WalletScheme = artifacts.require("./WalletScheme.sol");
 const DxController = artifacts.require("./DxController.sol");
 const DxAvatar = artifacts.require("./DxAvatar.sol");
